// web/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Language = {
  PT: 'PT',
  EN: 'EN'
} as const

export type Language = (typeof Language)[keyof typeof Language]


export const MediaType = {
  IMAGE_URL: 'IMAGE_URL',
  IMAGE_FILEID: 'IMAGE_FILEID',
  VIDEO_URL: 'VIDEO_URL',
  VIDEO_FILEID: 'VIDEO_FILEID',
  VIDEO_LOCAL: 'VIDEO_LOCAL',
  IMAGE_LOCAL: 'IMAGE_LOCAL'
} as const

export type MediaType = (typeof MediaType)[keyof typeof MediaType]


export const ProfileType = {
  SUPREME: 'SUPREME',
  SUPER_ADMIN: 'SUPER_ADMIN',
  ADMIN: 'ADMIN',
  MODERATOR: 'MODERATOR',
  USER: 'USER',
  BANNED: 'BANNED'
} as const

export type ProfileType = (typeof ProfileType)[keyof typeof ProfileType]


export const SourceType = {
  ANIME: 'ANIME',
  GAME: 'GAME',
  MANGA: 'MANGA',
  MOVIE: 'MOVIE'
} as const

export type SourceType = (typeof SourceType)[keyof typeof SourceType]